server/src/requirements.txt
alembic==0.8.4
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.22
blinker==1.4
boto==2.39.0
braintree==3.25.0
celery==3.1.20
contextlib2==0.5.1
decorator==4.0.9
docopt==0.4.0
elasticsearch==2.2.0
elasticsearch-dsl==2.0.0
Flask==0.10.1
Flask-Elasticsearch==0.2.5
Flask-Images==2.1.2
Flask-Login==0.3.2
Flask-Migrate==1.7.0
Flask-Script==2.0.5
Flask-Session==0.2.3
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
fudge==1.1.0
gnureadline==6.3.3
gunicorn==19.4.5
inflect==0.2.5
ipython==4.1.1
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.33
mailchimp==2.0.9
Mako==1.0.3
mandrill==1.0.57
MarkupSafe==0.23
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
pillowcase==2.0.0
psycopg2==2.6.1
ptyprocess==0.5.1
python-dateutil==2.4.2
python-editor==0.5
pytz==2015.7
raven==5.11.1
redis==2.10.5
requests==2.9.1
simplegeneric==0.8.1
simplejson==3.8.1
six==1.10.0
soundcloud==0.5.0
SQLAlchemy==1.0.11
traitlets==4.1.0
urllib3==1.14
Werkzeug==0.11.4
WTForms==2.1
